Complete redo of the seq id resolution algorithm using the new OSLT function; adjustments due to changes in the DB staructure; enable working with and without LMDB cache; adopt changes in the protobuf spec; minor bug fixes; minor improvements. JIRA: CXX-10335, CXX-10352, CXX-10353, CXX-10362
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@84834 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/objtools/pubseq_gateway/impl/cassandra/bioseq_info.hpp
@@ -77,11 +77,11 @@ CRequestStatus::ECode
 FetchCanonicalSeqId(shared_ptr<CCassConnection>  conn,
                     const string &  keyspace,
                     const string &  sec_seq_id,
-                    int  sec_seq_id_type,
+                    int16_t  sec_seq_id_type,
                     bool  sec_seq_id_type_provided,
                     string &  accession,                // output
-                    int &  version,                     // output
-                    int &  seq_id_type);                // output
+                    int16_t &  version,                 // output
+                    int16_t &  seq_id_type);            // output
 
 // e200_Ok: exactly one record found
 // e404_NotFound: no records found
